+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(mailbox) allow to reset password on mailboxes #834
+
 ## [1.16.0] - 2025-05-05
 
 ### Added